Moved modal to components directory and added option to change hash to add card
