--- client/components/hello-world.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export default function HelloWorld(props) {
-  return (
-    <h1>Hello, World!</h1>
-  );
-}
